Refactored to put main code in a class - just running from test
